IMPALA-2548: Codegen Tuple::MaterializeExprs() and use in TopN node
For the following benchmark query:
 select count(*) from (select l_orderkey from biglineitem order by l_orderkey limit 1000) a

The overall query time goes from 2.74s to 1.74s, with the top-n node
time going from 2.2s to 1.0s. There is no effect on sort node time.

The overall approach of this patch is to move the
TopNNode::InsertTupleRow() call into a cross-compiled batched function
(InsertBatch()), and then replace the MaterializeExprs() calls with
new functions built using the IRBuilder. This involves new codegen
utilities, such as CodegenAnyVal::WriteToSlot() and the ability to
hardcode in a MemPool pointer from which to make varlen data
allocations. This patch also adds a new timer measuring the time spent
inserting tuple rows.

The existing TestQueries::test_top_n and TestQueries::test_sort tests
pass with this patch.

Change-Id: Ib422a8d50303c21c6a228675157bf867e8619444
Reviewed-on: http://gerrit.cloudera.org:8080/1901
Reviewed-by: Skye Wanderman-Milne <skye@cloudera.com>
Tested-by: Internal Jenkins