IMPALA-3572: FE unit test coverage report with Jacoco.
This patch integrates Jacoco into Impala's FE test runs
for getting a code coverage report.

The instrumentation and reporting functionality is disabled
by default, and must be enabled explicitly, e.g., like this:
mvn test -DcodeCoverage

The code coverage report is stored in this location:
$IMPALA_HOME/logs/fe_tests/coverage

With additional changes, Jacoco can also be used to get code
coverage reports for our end-to-end tests, but that is left
for future work.

Change-Id: Id5e4f1b8afb91210d40622aadd3d21d7ed94c2a7
Reviewed-on: http://gerrit.cloudera.org:8080/3151
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins