IMPALA-3537: Move useHiveColLabels out of GlobalState and into Analyzer.
The bug: The useHiveColLabels flag was set in the Analyzer's global
state, so changes to the labelling scheme incorrectly leaked into
other query blocks, e.g., ancestor blocks. This could affect the colunm
labels of the final query results.

The flag really is a property of a specific query block, and not all
query blocks, so the fix is to make the flag a member of Analyzer.

Change-Id: Ie1e446474e1862c78dffd5192a175e043a1d0e90
Reviewed-on: http://gerrit.cloudera.org:8080/3133
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins