IMPALA-3143: Fix permission check when user belongs to supergroup
This commit fixes an issue where a permission check warning is
incorrectly generated when a user that belongs to the supergroup issues
a create database statement. The fix is to also check if the user
belongs to the supergroup when checking for permissions and ACLs.

Testing:
This change was tested manually by creating an HDFS directory D that
belonged to user/group A and running a CREATE DATABASE foo LOCALTION 'D'
statement as a user B that belongs to the 'supergroup' group; the test
verified that when the supergroup is taken into account, no permision
check warning is issued.

Change-Id: I2bb703ac70280c62db404db2c07d87139aa5ae0d
Reviewed-on: http://gerrit.cloudera.org:8080/3164
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins