Disable Kudu and remove startup option to enable it
Kudu is disabled for the 2.6.X release because stability and quality are
not production ready. People who want to try Kudu are expected to use a
different release called "Impala_Kudu".

For consistency, both the runtime flag is removed and the build variable
is disabled. The build variable is needed so the Kudu test service
(mini-cluster) and tests won't run. The runtime flag is removed because
disabling the Kudu build means a stubbed Kudu client library will be
used instead of the real library so the Kudu backend won't work even if
the flag existed and was enabled.

Change-Id: Ie2c417c8dcbd1bd74d47d1e8294698992ee36eba
Reviewed-on: http://gerrit.cloudera.org:8080/3204
Reviewed-by: Marcel Kornacker <marcel@cloudera.com>
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: Internal Jenkins