IMPALA-3600: Add missing admission control tests
* -require_username (not strictly admission control related
  but it came up in the context of RM).
* Coverage of failure cases: The handling of the full queue
  case wasn't being verified. This changes existing stress
  test to expect a specific message when the queue is full.
* Requesting MAXINT memory, which previously led to an
  overflow in the pool-level mem tracker accounting.

This does not yet address:
* Changing pool cfg while running
* Verify profile string for queued reason

This is just a minimal incremental change to get additional
coverage. Right now, many of the tests rely on some
pre-defined configuration files which is cumbersome. In the
future, we plan on refreshing the configuration story at
which point we should also build more general test
infrastructure for easily testing different configurations.

Change-Id: I6682b15a5caac5824384c4b48a7b40afa2548954
Reviewed-on: http://gerrit.cloudera.org:8080/3272
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Matthew Jacobs <mj@cloudera.com>