IMPALA-3664: S3A test_keys_do_not_work fails
Previously when we created a new connection to S3 via the S3A
connector, the S3AFileSystem object was created only then. However,
recently it was realized that the object could be created even before
we created a connection to the S3A file system; so the configuration
values we pass through to the connection would be ignored.

This is fixed by forcing the FS builder to return a new instance of
the filesystem object, which will read the configuration changes we
make.

The test is also split up to test a DDL statement in a separate test
because we would want that statement to succeed as it goes through the
HMS and the SELECT statement to fail.

Change-Id: I23b541eef747dd62e59390f8cc9ac6e5742ead40
Reviewed-on: http://gerrit.cloudera.org:8080/3392
Reviewed-by: Sailesh Mukil <sailesh@cloudera.com>
Tested-by: Sailesh Mukil <sailesh@cloudera.com>